test: factor out util/admin package for setup of ceph admin client
